+++ workspaces/servicenow/plugins/servicenow/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage-community/plugin-servicenow
 
+## 1.15.2
+
+### Patch Changes
+
+- 1b590ba: Updated dependency `@playwright/test` to `1.63.0`.
+- 3b2eddd: Backstage version bump to v1.54.7
+- Updated dependencies [3b2eddd]
+  - @backstage-community/plugin-servicenow-common@1.13.1
+
+## 1.15.1
+
+### Patch Changes
+
+- 4813d14: Updated dependency `@playwright/test` to `1.62.1`.
+
 ## 1.15.0
 
 ### Minor Changes